Move parallel part into a separate file
